+++ packages/imperative/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Masked sensitive command-line options that are supplied in the equals-separated form (for example, `--user=example` or `-u=example`) before the command line is written to the log, and also mask secure values that contain embedded whitespace.
+
 ## `5.27.19`
 
 - BugFix: Added the `base64EncodedAuth` session property to the list of session properties redacted from Imperative debug logs. [#2780](https://github.com/zowe/zowe-cli/pull/2780)